+++ src/core/text-rendering/sdf/index.ts
@@ -17,4 +17,4 @@
  * limitations under the License.
  */
 
-export * from './Utils.js';
+// export * from './Utils.js';
